infra/pipecat-agent/requirements.txt
pipecat-ai[livekit,deepgram,cartesia,openai,silero,azure,google,elevenlabs,websocket]
livekit-api
fastapi
python-multipart
uvicorn[standard]
loguru
python-dotenv
twilio
